Move StreamKind to its own file / submodule
